create operations directory and move world_info.rb
